+++ lib/src/clojure_lsp/feature/diagnostics.clj
@@ -257,10 +257,20 @@
                                (remove exclude-def?))
           var-nses (set (map :ns var-definitions)) ;; optimization to limit usages to internal namespaces, or in the case of a single file, to its namespaces
           all-dependents (q/nses-and-dependents-analysis project-db var-nses)
-          ignore-test-references? (get-in kondo-config [:linters :clojure-lsp/unused-public-var :ignore-test-references?] false)
+          ignore-test-references? (get-in kondo-config
+                                          [:linters :clojure-lsp/unused-public-var :ignore-test-references?]
+                                          false)
+          test-locations-regex (into #{}
+                                     (map re-pattern
+                                          (settings/get project-db
+                                                        [:test-locations-regex]
+                                                        shared/test-locations-regex-default)))
+          dependents-without-tests (into {}
+                                         (remove (fn [[uri _]]
+                                                   (some #(re-find % uri) test-locations-regex))
+                                                 all-dependents))
           dependents (if ignore-test-references?
-                       (into {}
-                             (remove (fn [[uri _]] (string/ends-with? uri "_test.clj")) all-dependents))
+                       dependents-without-tests
                        all-dependents)
           var-usages (into #{}
                            (comp